Fixes payload serialization

+++ src/payload.c
@@ -154,7 +154,7 @@ int payload_serialize(payload *p, uint8_t *buffer, size_t buffer_size, size_t *b
 
     log_debug("Adds receiver number");
     utils_uint64_to_be(buffer + ln, p->receiver);
-    ln += sizeof(uint64_t);
+    ln += sizeof(uint32_t);
 
     log_debug("Adds payload number");
     utils_uint64_to_be(buffer + ln, p->number);
